--- Maple2.Model/Enum/ShopBuyDay.cs
@@ -1,11 +1,0 @@
-﻿namespace Maple2.Model.Enum;
-
-public enum ShopBuyDay : byte {
-    Sunday = 1,
-    Monday = 2,
-    Tuesday = 3,
-    Wednesday = 4,
-    Thursday = 5,
-    Friday = 6,
-    Saturday = 7,
-}
