--- Maple2.Model/Enum/ShopCurrencyType.cs
@@ -1,19 +1,0 @@
-﻿namespace Maple2.Model.Enum;
-
-public enum ShopCurrencyType : byte {
-    Meso = 0,
-    Item = 1,
-    ValorToken = 2,
-    Treva = 3,
-    Meret = 4,
-    Rue = 5,
-    HaviFruit = 6,
-    GuildCoin = 7,
-    ReverseCoin = 8,
-    EventMeret = 9,
-    GameMeret = 10, // RedMeret
-    MentorToken = 11,
-    MenteeToken = 12,
-    StarPoint = 13,
-    MesoToken = 14,
-}
