--- Maple2.Model/Enum/ShopFrameType.cs
@@ -1,9 +1,0 @@
-﻿namespace Maple2.Model.Enum;
-
-public enum ShopFrameType : byte {
-    Default = 0,
-    Unknown = 1,
-    Star = 2,
-    StyleCrate = 3,
-    Capsule = 4,
-}
