--- Maple2.Model/Enum/ShopItemLabel.cs
@@ -1,9 +1,0 @@
-﻿namespace Maple2.Model.Enum;
-
-public enum ShopItemLabel : byte {
-    None = 0,
-    New = 1,
-    Event = 2,
-    HalfPrice = 3,
-    Special = 4,
-}
