--- Maple2.Model/Enum/ShopRestockInterval.cs
@@ -1,8 +1,0 @@
-﻿namespace Maple2.Model.Enum;
-
-public enum ShopRestockInterval : byte {
-    Minute = 0,
-    Day = 1,
-    Week = 2,
-    Month = 3,
-}
