--- Maple2.Model/Enum/ShopSkin.cs
@@ -1,9 +1,0 @@
-﻿namespace Maple2.Model.Enum;
-
-public enum ShopSkin : byte {
-    Default = 0,
-    Unknown = 1,
-    Star = 2,
-    StyleCrate = 3,
-    Capsule = 4,
-}
